+++ CHANGES.rst
@@ -12,6 +12,7 @@ in progress
     - Apply custom layout settings based on poster name
     - Add matrix-based chooser for language vs. name,
       see. https://examples.posterkit.net/lqdn-gafam-campaign/chooser.html
+    - Dynamic refitting if body content overflows its container element
 
 2018-01-15 0.0.0
 ----------------